Move refreshService tests to their own section.
