+++ test/vc-data-model-1.0/20-advanced.js
@@ -75,52 +75,6 @@ describe('Advanced Documents', function() {
     });
   });
 
-  // https://w3c.github.io/vc-data-model/#refreshing
-  describe('Refreshing', function() {
-    it('`refreshService` MUST provide one or more refresh services', async function() {
-      // test that `refreshService` is either an array or an object
-      const doc = await util.generate('example-011.jsonld', generatorOptions);
-      const isArray = Array.isArray(doc.refreshService) &&
-        doc.refreshService.length > 0;
-      const isObject = doc.refreshService && typeof doc.refreshService.id === 'string';
-      expect(isArray || isObject).to.be.true;
-    });
-
-    describe('each object within `refreshService`...', function() {
-      // if there are multiple objects, loop these tests
-      it('MUST specify a `type` property with a valid value', async function() {
-        // test for `type` property existence
-        const doc = await util.generate('example-011.jsonld', generatorOptions);
-        const services = [].concat(doc.refreshService);
-
-        for(let service of services) {
-          expect(service.type).to.be.a('string');
-        }
-      });
-
-      it('MUST specify an `id` property', async function() {
-        // test for `serviceEndpoint` property existence
-        const doc = await util.generate('example-011.jsonld', generatorOptions);
-        const services = [].concat(doc.refreshService);
-
-        for(let service of services) {
-          expect(service.id).to.be.a('string');
-        }
-      });
-
-      it('value of `id` MUST be a URL identifying a service endpoint', async function() {
-        // test that the `id`'s value is a valid URL
-        // (possibly) attempt to dereference the service endpoint to validate it's "locator-ness" (i.e. URL vs. URI)
-        const doc = await util.generate('example-011.jsonld', generatorOptions);
-        const services = [].concat(doc.refreshService);
-
-        for(let service of services) {
-          expect(service.id).to.match(uriRegex);
-        }
-      });
-    });
-  });
-
   // https://w3c.github.io/vc-data-model/#evidence
   describe('Evidence', function() {
     // `evidence` is optional, so these should only be run if that term is present
@@ -149,19 +103,4 @@ describe('Advanced Documents', function() {
     });
   });
 
-  // https://w3c.github.io/vc-data-model/#zero-knowledge-proofs
-  // ...see dedicated ZKP test file.
-
-  // Subject is the Holder is at risk (and vague); so skipping (for now?)
-  // https://w3c.github.io/vc-data-model/#subject-is-the-holder
-
-  // Disputes is currently under-defined and (consequently) untestable
-  // https://w3c.github.io/vc-data-model/#disputes
-
-  // Authorization contains a "specification use" requirement--which cannot be
-  // tested in code; so skipping.
-  // https://w3c.github.io/vc-data-model/#authorization
-
-  // Syntactic Sugar is a descriptive (only) section; so skipping.
-  // https://w3c.github.io/vc-data-model/#syntactic-sugar
 });
